backend/requirements.txt
# ============================
# Backend Requirements
# Python 3.10+, CUDA GPU required
# ============================

# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0
sse-starlette==2.1.0

# Computer Vision & Detection
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37,<5         # Florence-2 / Grounding DINO via HuggingFace
timm>=0.9.16                   # Backbone models
supervision==0.20.0            # Detection utilities & visualization
Pillow>=10.2.0
opencv-python-headless>=4.9.0
numpy>=1.26.4

# ByteTrack (installed via local clone or pip)
# pip install git+https://github.com/ifzhang/ByteTrack.git
# OR via requirements below if packaged:
lap>=0.5.13                     # Linear Assignment Problem solver for ByteTrack
scipy>=1.12.0
filterpy==1.4.5

# Async & Task Queue
asyncio-mqtt==0.16.2           # Optional MQTT (not required by default)
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
loguru==0.7.2
tqdm>=4.66.2

# ZIP and file handling
zipfile36==0.1.3               # Extended zipfile (stdlib zipfile sufficient too)

# Optional: Redis/Celery for task queue scaling
# redis==5.0.4
# celery==5.3.6
